requirements_1.txt
nsepython==2.9
requests==2.31.0
pytz==2024.1
python-dateutil==2.9.0
